Fix race condition in IsValidationPending, apply minor review fixes
