+++ src/Components/Web/test/Forms/EditFormTest.cs
@@ -204,6 +204,188 @@ public class EditFormTest
         Assert.Equal("some value", editFormAttributes["custom attribute"]);
     }
 
+    [Fact]
+    public async Task Submit_AwaitsAsyncValidationBeforeOnValidSubmit()
+    {
+        var editContext = new EditContext(new TestModel());
+        var field = editContext.Field(nameof(TestModel.StringProperty));
+        TestAsyncValidator validator = null;
+        var validSubmitCount = 0;
+        var rootComponent = new AsyncEditFormHostComponent
+        {
+            EditContext = editContext,
+            Configure = current =>
+            {
+                current.Configure(field, new ValidationConfig { Outcome = ValidationOutcome.Valid });
+                current.GetGate(field);
+            },
+            Created = current => validator = current,
+            OnValidSubmit = _ => validSubmitCount++,
+        };
+        await RenderAsyncRootAsync(rootComponent);
+
+        var dispatchTask = _testRenderer.DispatchEventAsync(GetSubmitEventHandlerId(), EventArgs.Empty);
+        await WaitUntilAsync(() => validator.FormValidationStartCount == 1);
+
+        Assert.Equal(0, validSubmitCount);
+
+        validator.OpenGate(field, ValidationOutcome.Valid);
+        await dispatchTask.WaitAsync(DefaultAsyncTimeout);
+
+        Assert.Equal(1, validSubmitCount);
+    }
+
+    [Fact]
+    public async Task Submit_InvalidAsyncValidation_FiresOnInvalidSubmit()
+    {
+        var editContext = new EditContext(new TestModel());
+        var field = editContext.Field(nameof(TestModel.StringProperty));
+        var validSubmitCount = 0;
+        var invalidSubmitCount = 0;
+        var rootComponent = new AsyncEditFormHostComponent
+        {
+            EditContext = editContext,
+            Configure = current => current.Configure(field, new ValidationConfig { Outcome = ValidationOutcome.Invalid, ErrorMessage = "Invalid" }),
+            OnValidSubmit = _ => validSubmitCount++,
+            OnInvalidSubmit = _ => invalidSubmitCount++,
+        };
+        await RenderAsyncRootAsync(rootComponent);
+
+        await _testRenderer.DispatchEventAsync(GetSubmitEventHandlerId(), EventArgs.Empty).WaitAsync(DefaultAsyncTimeout);
+
+        Assert.Equal(0, validSubmitCount);
+        Assert.Equal(1, invalidSubmitCount);
+        Assert.Equal(new[] { "Invalid" }, editContext.GetValidationMessages(field));
+    }
+
+    [Fact]
+    public async Task Submit_AsyncValidatorThrows_FiresOnInvalidSubmitWithFaultedContext()
+    {
+        var editContext = new EditContext(new TestModel());
+        var field = editContext.Field(nameof(TestModel.StringProperty));
+        var validSubmitCount = 0;
+        var invalidSubmitCount = 0;
+        var observedFaulted = false;
+        var rootComponent = new AsyncEditFormHostComponent
+        {
+            EditContext = editContext,
+            Configure = current => current.Configure(field, new ValidationConfig { Outcome = ValidationOutcome.ThrowInfraException }),
+            OnValidSubmit = _ => validSubmitCount++,
+            OnInvalidSubmit = context =>
+            {
+                invalidSubmitCount++;
+                observedFaulted = context.IsValidationFaulted();
+            },
+        };
+        await RenderAsyncRootAsync(rootComponent);
+
+        await _testRenderer.DispatchEventAsync(GetSubmitEventHandlerId(), EventArgs.Empty).WaitAsync(DefaultAsyncTimeout);
+
+        Assert.Equal(0, validSubmitCount);
+        Assert.Equal(1, invalidSubmitCount);
+        Assert.True(observedFaulted);
+    }
+
+    [Fact]
+    public async Task Submit_WithPendingFieldTask_CancelsFieldTaskAndRunsFormValidation()
+    {
+        var editContext = new EditContext(new TestModel());
+        var field = editContext.Field(nameof(TestModel.StringProperty));
+        TestAsyncValidator validator = null;
+        var validSubmitCount = 0;
+        var rootComponent = new AsyncEditFormHostComponent
+        {
+            EditContext = editContext,
+            Configure = current => current.Configure(field, new ValidationConfig { Outcome = ValidationOutcome.Valid }),
+            Created = current => validator = current,
+            OnValidSubmit = _ => validSubmitCount++,
+        };
+        await RenderAsyncRootAsync(rootComponent);
+        var pendingCts = new CancellationTokenSource();
+        var pendingTcs = new TaskCompletionSource(TaskCreationOptions.RunContinuationsAsynchronously);
+        using var pendingRegistration = pendingCts.Token.Register(() => pendingTcs.TrySetCanceled(pendingCts.Token));
+        editContext.AddValidationTask(field, pendingTcs.Task, pendingCts);
+
+        await _testRenderer.DispatchEventAsync(GetSubmitEventHandlerId(), EventArgs.Empty).WaitAsync(DefaultAsyncTimeout);
+
+        Assert.True(pendingCts.IsCancellationRequested);
+        Assert.False(editContext.IsValidationPending(field));
+        Assert.Equal(1, validator.FormValidationStartCount);
+        Assert.Equal(1, validSubmitCount);
+    }
+
+    private async Task RenderAsyncRootAsync(AsyncEditFormHostComponent rootComponent)
+    {
+        var componentId = _testRenderer.AssignRootComponentId(rootComponent);
+        await _testRenderer.RenderRootComponentAsync(componentId);
+    }
+
+    private ulong GetSubmitEventHandlerId()
+    {
+        var editFormComponentId = _testRenderer.Batches.Last().ReferenceFrames.AsEnumerable()
+            .Where(frame => frame.FrameType == RenderTreeFrameType.Component)
+            .Where(frame => frame.Component is EditForm)
+            .Select(frame => frame.ComponentId)
+            .Single();
+        var editFormFrames = _testRenderer.GetCurrentRenderTreeFrames(editFormComponentId);
+        return editFormFrames.AsEnumerable()
+            .Where(frame => frame.FrameType == RenderTreeFrameType.Attribute)
+            .Where(frame => frame.AttributeName == "onsubmit")
+            .Select(frame => frame.AttributeEventHandlerId)
+            .Single();
+    }
+
+    private static async Task WaitUntilAsync(Func<bool> condition)
+    {
+        var start = DateTime.UtcNow;
+        while (!condition())
+        {
+            if (DateTime.UtcNow - start > DefaultAsyncTimeout)
+            {
+                throw new TimeoutException("The expected condition was not reached before the timeout.");
+            }
+
+            await Task.Yield();
+        }
+    }
+
+    private static readonly TimeSpan DefaultAsyncTimeout = TimeSpan.FromSeconds(5);
+
+    private sealed class AsyncEditFormHostComponent : AutoRenderComponent
+    {
+        public EditContext EditContext { get; set; }
+
+        public Action<TestAsyncValidator> Configure { get; set; }
+
+        public Action<TestAsyncValidator> Created { get; set; }
+
+        public Action<EditContext> OnValidSubmit { get; set; }
+
+        public Action<EditContext> OnInvalidSubmit { get; set; }
+
+        protected override void BuildRenderTree(RenderTreeBuilder builder)
+        {
+            builder.OpenComponent<EditForm>(0);
+            builder.AddComponentParameter(1, nameof(EditForm.EditContext), EditContext);
+            if (OnValidSubmit is not null)
+            {
+                builder.AddComponentParameter(2, nameof(EditForm.OnValidSubmit), EventCallback.Factory.Create(this, OnValidSubmit));
+            }
+            if (OnInvalidSubmit is not null)
+            {
+                builder.AddComponentParameter(3, nameof(EditForm.OnInvalidSubmit), EventCallback.Factory.Create(this, OnInvalidSubmit));
+            }
+            builder.AddComponentParameter(4, nameof(EditForm.ChildContent), (RenderFragment<EditContext>)(context => childBuilder =>
+            {
+                childBuilder.OpenComponent<TestAsyncValidatorComponent>(0);
+                childBuilder.AddComponentParameter(1, nameof(TestAsyncValidatorComponent.Configure), Configure);
+                childBuilder.AddComponentParameter(2, nameof(TestAsyncValidatorComponent.Created), EventCallback.Factory.Create<TestAsyncValidator>(this, Created));
+                childBuilder.CloseComponent();
+            }));
+            builder.CloseComponent();
+        }
+    }
+
     private static EditForm FindEditFormComponent(CapturedBatch batch)
         => batch.ReferenceFrames
                 .Where(f => f.FrameType == RenderTreeFrameType.Component)
